requirements.txt
astroid==2.4.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
facebook-sdk==3.1.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1